Remove invalid functions for folded and literal encoding of strings + Move Yaml encoding tests to shared
